requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
lxml==4.9.3
Pillow==10.0.0
requests==2.31.0
telebot==0.0.5
pyTelegramBotAPI==4.11.0